separate 2d and 3d
