--- project/CMakeLists.txt
@@ -1,4 +1,0 @@
-# Add an executable
-add_executable(FilletMain main.cpp)
-
-target_link_libraries(FilletMain manifold)
