+++ scripts/visualizer-polygon.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 from typing import List, Dict, Any
-
+import sys
 
 
 def plot_polygon(ax, points_list, label, color, closed=True, linestyle='-', marker='o', markersize=5, linewidth=1.5, show_indices=False, index_offset=0.2):
@@ -134,7 +134,7 @@ def read_polygon_tests_file(filename: str) -> List[Dict[str, Any]]:
 if __name__ == "__main__":
     fig, (ax1, ax2) = plt.subplots(2, 1, figsize=(15, 20))
 
-    result = read_polygon_tests_file("result.txt")
+    result = read_polygon_tests_file(sys.argv[1])
     
     Poly = [
     [0, 0] ,[0, 5] ,[5, 5] ,[5, 0] ,[1, 4.5] ,[0.975528, 4.65451] ,[0.904508, 4.79389] ,[0.793893, 4.90451] ,[0.654508, 4.97553] ,[0.5, 5] ,[0.345492, 4.97553] ,[0.206107, 4.90451] ,[0.0954915, 4.79389] ,[0.0244717, 4.65451] ,[0, 4.5] ,[0.0244717, 4.34549] ,[0.0954915, 4.20611] ,[0.206107, 4.09549] ,[0.345492, 4.02447] ,[0.5, 4] ,[0.654508, 4.02447] ,[0.793893, 4.09549] ,[0.904508, 4.20611] ,[0.975528, 4.34549] ,[1, 4.5] ,[0.975528, 4.65451] ,[0.904508, 4.79389] ,[0.793893, 4.90451] ,[0.654508, 4.97553] ,[0.5, 5] ,[0.345492, 4.97553] ,[0.206107, 4.90451] ,[0.0954915, 4.79389] ,[0.0244717, 4.65451] ,[0, 4.5] ,[0.0244717, 4.34549] ,[0.0954915, 4.20611] ,[0.206107, 4.09549] ,[0.345492, 4.02447] ,[0.5, 4] ,[0.654508, 4.02447] ,[0.793893, 4.09549] ,[0.904508, 4.20611] ,[0.975528, 4.34549] ,[5, 4.5] ,[4.97553, 4.65451] ,[4.90451, 4.79389] ,[4.79389, 4.90451] ,[4.65451, 4.97553] ,[4.5, 5] ,[4.34549, 4.97553] ,[4.20611, 4.90451] ,[4.09549, 4.79389] ,[4.02447, 4.65451] ,[4, 4.5] ,[4.02447, 4.34549] ,[4.09549, 4.20611] ,[4.20611, 4.09549] ,[4.34549, 4.02447] ,[4.5, 4] ,[4.65451, 4.02447] ,[4.79389, 4.09549] ,[4.90451, 4.20611] ,[4.97553, 4.34549] ,[5, 4.5] ,[4.97553, 4.65451] ,[4.90451, 4.79389] ,[4.79389, 4.90451] ,[4.65451, 4.97553] ,[4.5, 5] ,[4.34549, 4.97553] ,[4.20611, 4.90451] ,[4.09549, 4.79389] ,[4.02447, 4.65451] ,[4, 4.5] ,[4.02447, 4.34549] ,[4.09549, 4.20611] ,[4.20611, 4.09549] ,[4.34549, 4.02447] ,[4.5, 4] ,[4.65451, 4.02447] ,[4.79389, 4.09549] ,[4.90451, 4.20611] ,[4.97553, 4.34549] ,[5, 0.5] ,[4.97553, 0.654508] ,[4.90451, 0.793893] ,[4.79389, 0.904508] ,[4.65451, 0.975528] ,[4.5, 1] ,[4.34549, 0.975528] ,[4.20611, 0.904508] ,[4.09549, 0.793893] ,[4.02447, 0.654508] ,[4, 0.5] ,[4.02447, 0.345492] ,[4.09549, 0.206107] ,[4.20611, 0.0954915] ,[4.34549, 0.0244717] ,[4.5, 1.11022e-16] ,[4.65451, 0.0244717] ,[4.79389, 0.0954915] ,[4.90451, 0.206107] ,[4.97553, 0.345492] ,[5, 0.5] ,[4.97553, 0.654508] ,[4.90451, 0.793893] ,[4.79389, 0.904508] ,[4.65451, 0.975528] ,[4.5, 1] ,[4.34549, 0.975528] ,[4.20611, 0.904508] ,[4.09549, 0.793893] ,[4.02447, 0.654508] ,[4, 0.5] ,[4.02447, 0.345492] ,[4.09549, 0.206107] ,[4.20611, 0.0954915] ,[4.34549, 0.0244717] ,[4.5, 1.11022e-16] ,[4.65451, 0.0244717] ,[4.79389, 0.0954915] ,[4.90451, 0.206107] ,[4.97553, 0.345492] ,[1, 0.5] ,[0.975528, 0.654508] ,[0.904508, 0.793893] ,[0.793893, 0.904508] ,[0.654508, 0.975528] ,[0.5, 1] ,[0.345492, 0.975528] ,[0.206107, 0.904508] ,[0.0954915, 0.793893] ,[0.0244717, 0.654508] ,[1.11022e-16, 0.5] ,[0.0244717, 0.345492] ,[0.0954915, 0.206107] ,[0.206107, 0.0954915] ,[0.345492, 0.0244717] ,[0.5, 0] ,[0.654508, 0.0244717] ,[0.793893, 0.0954915] ,[0.904508, 0.206107] ,[0.975528, 0.345492] ,[1, 0.5] ,[0.975528, 0.654508] ,[0.904508, 0.793893] ,[0.793893, 0.904508] ,[0.654508, 0.975528] ,[0.5, 1] ,[0.345492, 0.975528] ,[0.206107, 0.904508] ,[0.0954915, 0.793893] ,[0.0244717, 0.654508] ,[1.11022e-16, 0.5] ,[0.0244717, 0.345492] ,[0.0954915, 0.206107] ,[0.206107, 0.0954915] ,[0.345492, 0.0244717] ,[0.5, 0] ,[0.654508, 0.0244717] ,[0.793893, 0.0954915] ,[0.904508, 0.206107] ,[0.975528, 0.345492]
