+++ src/CMakeLists.txt
@@ -27,7 +27,7 @@ set(
   quickhull.cpp
   sdf.cpp
   smoothing.cpp
-  fillet.cpp
+  fillet2d.cpp
   sort.cpp
   subdivision.cpp
   tree2d.cpp
